refactor: split cli_compare_fold without joining the CLI import-cycle SCC
Codex review: the previous split put _fold_evidence_depth_into_json (which
reaches cli_dump_helpers) in the new module, forcing an IMPORT_CYCLE_ALLOWLIST
extension — against AGENTS.md's explicit rule that a new allowlist member
needs an ADR/maintainer sign-off, not a routine split. Keeps that function in
cli_compare_helpers.py and moves only _resolve_scoped_gate_findings/
_fold_scoped_compat_into_text/_SEVERITY_TO_SUMMARY_BUCKET, which reach only
leaf report-formatting modules (reporter, reporter_markdown, severity,
checker_policy) -- cli_compare_fold no longer joins the SCC at all, so the
allowlist change is reverted.
